renamed header to navbar, renamed navbar to authentication for clarity
