+++ debug.log
@@ -128,3 +128,8 @@
 [1110/132803.416:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [1110/132803.416:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [1110/132803.416:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1110/133253.359:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1110/133253.359:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1110/133253.359:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1110/133253.360:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1110/133253.360:ERROR:crash_report_database_win.cc(469)] failed to stat report
