+++ src/components/Authentication/Authentication.css
@@ -1,0 +1,98 @@
+
+.login{
+    animation-duration: 20s;
+    animation-delay: 10s;
+    animation-iteration-count: infinite;
+  }
+
+h1 {
+    color: white;
+}
+h2{
+    color: white;
+}
+
+button {
+    padding: 20px;
+    border: #00d8ff;
+    font-family: fantasy;
+    font-size: 25px;
+    box-shadow: 0 3px 6px #00d8ff, 0 3px 6px #00d8ff; 
+    background: white;
+}
+
+button:hover {
+    background:#00d8ff;    
+    
+}
+
+  body{
+    font-family: 'Roboto', sans-serif;
+  }
+    .container{
+      width:250px;
+      height:250px;
+      margin:20px auto 0 auto;
+      background:#222;
+  
+      display:flex;
+      align-items: center;
+      justify-content: center;
+    }
+  
+    .react-logo{
+      display: inline-block;
+      border-radius:50%;
+      border:10px solid #00d8ff;
+  
+      animation-name:rotate;
+      animation-duration: 5s;
+      animation-iteration-count: infinite;
+      animation-timing-function: linear;
+  
+    }
+  
+    .react-logo , .react-logo:before , .react-logo:after{
+      position:absolute;
+      display: inline-block;
+      width:200px;
+      height:70px;
+      border-radius:50%;
+      border:10px solid #00d8ff;
+    }
+  
+    .nucleo , .react-logo , .react-logo:before , .react-logo:after {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+    }
+  
+    .nucleo{
+      width:40px;
+      height:40px;
+      border-radius:100%;
+      background:#00d8ff;
+    }
+  
+    .react-logo:before , .react-logo:after{
+      content:"";
+    }
+  
+    .react-logo:after{
+      transform:rotate(-57deg);
+    }
+  
+    .react-logo:before{
+      transform:rotate(57deg);
+    }
+
+  
+    @keyframes rotate{
+  
+      100%{
+        transform:rotate(360deg);
+      }
+    }
+
+
+
