+++ src/pages/Login.js
@@ -4,8 +4,8 @@ import { useAuth0 } from "../react-auth0-spa";
 import { Link } from "react-router-dom";
 import styles from "../styles"
 import Jumbotron from "../components/Jumbotron/Jumbo";
+import Authentication from "../components/Authentication/Authentication"
 import NavBar from "../components/NavBar/NavBar"
-import Header from "../components/Header/Header"
 
 export function Login ({ children }) {
   const { isAuthenticated, loginWithRedirect, logout } = useAuth0();
@@ -16,12 +16,12 @@ export function Login ({ children }) {
      
       {!isAuthenticated && (
           <span>
-           <NavBar/>
+           <Authentication/>
           </span>               
       )}      
       {isAuthenticated && (
         <span>
-        <Header/>   
+        <NavBar/>   
         <div className="list-overflow-container">
            There will be a list of top/stickied discussions here
               <ul className="list-group">{children}</ul>
