+++ src/pages/Posts.js
@@ -4,7 +4,7 @@ import { useAuth0 } from "../react-auth0-spa";
 import { Link } from "react-router-dom";
 
 import Jumbotron from "../components/Jumbotron/Jumbo";
-import Header from "../components/Header/Header"
+import NavBar from "../components/NavBar/NavBar"
 
 export function Posts ({ children }) {
   const { isAuthenticated, loginWithRedirect, logout } = useAuth0();
@@ -20,7 +20,7 @@ export function Posts ({ children }) {
       )}      
       {isAuthenticated && (
         <span>
-          <Header />
+          <NavBar />
          <div className="list-overflow-container">
            There will be sortable list of discussions here
               <ul className="list-group">{children}</ul>
