+++ src/pages/Profile.js
@@ -2,7 +2,7 @@
 
 import React, { Fragment } from "react";
 import { useAuth0 } from "../react-auth0-spa";
-import Header from "../components/Header/Header"
+import NavBar from "../components/NavBar/NavBar"
 
 const Profile = () => {
   const { loading, user } = useAuth0();
@@ -15,7 +15,7 @@ const Profile = () => {
 
   return (
     <Fragment>
-      <Header />
+      <NavBar />
       <img src={user.picture} alt="Profile" />
       <h2>{user.name}</h2>
       <p>{user.email}</p>
